Aggiunta criptazione password

+++ Backend/Service-Platform/login-dispatcher-service/src/main/java/com/repository/LoginInfoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.model.entity.LoginInfo;
 
-public interface LoginInfoRepository extends JpaRepository<LoginInfo, Integer> {
+public interface LoginInfoRepository extends JpaRepository<LoginInfo, Long> {
 
 	LoginInfo findByEmail(String email);
